add Private ROute component
