+++ TicketManager.csproj
@@ -14,9 +14,9 @@
     </ItemGroup>
     <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.0.2" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.1" />
-        <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="3.0.1" />
-        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.0.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.2" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="3.0.2" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.0.2" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.0.1">
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
             <PrivateAssets>all</PrivateAssets>
@@ -24,7 +24,6 @@
         <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.0.1" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.0.1" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.0.1" />
-        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.0.1" />
         <PackageReference Include="Microsoft.OpenApi" Version="1.1.4" />
         <PackageReference Include="Moq" Version="4.13.1" />
         <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
