+++ client/src/App.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import { Router } from "react-router-dom";
-import * as createHistory from "history";
 import Layout from "./pages/Layout";
 import { useAuth0 } from "./authentication/auth0";
-
+// import history from "./utils/history";
+import * as createHistory from "history";
 export const history = createHistory.createBrowserHistory();
-
 export default function App() {
   const { loading } = useAuth0();
 
@@ -14,10 +13,8 @@ export default function App() {
   }
 
   return (
-    <>
-      <Router history={history}>
-        <Layout />
-      </Router>
-    </>
+    <Router history={history}>
+      <Layout />
+    </Router>
   );
 }
